data access layer implemented
Also created a User model with basic functionality like validation,
hashing password, serializing object to JSON and so on.

DataError exception added which represents error generic error related
to data layer, which is now not functioning
as expected because of how Node.js
error handling works with async functions.

Note: fixed security middleware to rely not on content-type,
but on whether passed data is JSON actually.